+++ providers/ic/src/providers/publish/publish.ic.ts
@@ -1,21 +1,32 @@
-import {Deck, Publish, PublishUrl} from '@deckdeckgo/editor';
+import {Deck, DeckPublish, Doc, DocPublish, PublishUrl} from '@deckdeckgo/editor';
 
 import {_SERVICE as StorageBucketActor} from '../../canisters/storage/storage.did';
 
+import {BucketActor} from '../../utils/manager.utils';
 import {publishDeck} from '../../utils/publish.deck.utils';
 import {uploadResources} from '../../utils/publish.resources.utils';
 import {publishOverview} from '../../utils/publish.overview.utils';
 import {getStorageActor} from '../../utils/storage.utils';
-import {BucketActor} from '../../utils/manager.utils';
+import {publishDoc} from '../../utils/publish.doc.utils';
 
-export const publish: Publish = async ({deck}: {deck: Deck; config: Record<string, string>}): Promise<Deck> => {
-  await uploadResources({deck});
+export const deckPublish: DeckPublish = async ({deck}: {deck: Deck; config: Record<string, string>}): Promise<Deck> => {
+  await uploadResources({meta: deck.data.meta});
 
-  const {storageUpload, deckPublishData, deck: updatedDeck} = await publishDeck({deck});
+  const {storageUpload, publishData, deck: updatedDeck} = await publishDeck({deck});
 
-  await publishOverview({storageUpload, deckPublishData, deckId: updatedDeck.id});
+  await publishOverview({storageUpload, publishData, dataId: updatedDeck.id});
 
   return updatedDeck;
+};
+
+export const docPublish: DocPublish = async ({doc}: {doc: Doc}): Promise<Doc> => {
+  await uploadResources({meta: doc.data.meta});
+
+  const {storageUpload, publishData, doc: updatedDoc} = await publishDoc({doc});
+
+  await publishOverview({storageUpload, publishData, dataId: updatedDoc.id});
+
+  return updatedDoc;
 };
 
 export const publishUrl: PublishUrl = async () => {