+++ providers/ic/stencil.config.ts
@@ -23,7 +23,7 @@ export const config: Config = {
       serviceWorker: null
     },
     {
-      type: 'dist-custom-elements-bundle'
+      type: 'dist-custom-elements'
     }
   ],
   plugins: [