+++ studio/src/app/components/editor/deck/actions/deck/app-actions-deck/app-actions-deck.tsx
@@ -8,12 +8,11 @@ import {ConnectionState, DeckdeckgoEventDeckRequest} from '@deckdeckgo/types';
 import {selectDeckSlide} from '@deckdeckgo/editor';
 
 import remoteStore from '../../../../../../stores/remote.store';
-import editorStore from '../../../../../../stores/editor.store';
-import userStore from '../../../../../../stores/user.store';
-import shareStore from '../../../../../../stores/share.store';
 import i18n from '../../../../../../stores/i18n.store';
 
 import {MoreAction} from '../../../../../../types/editor/more-action';
+
+import {share} from '../../../../../../utils/core/share.utils';
 
 import {AppIcon} from '../../../../../core/app-icon/app-icon';
 
@@ -47,6 +46,9 @@ export class AppActionsDeck {
 
   @Event()
   private stepTo: EventEmitter<HTMLElement | undefined>;
+
+  @Event()
+  private openEmbed: EventEmitter<void>;
 
   private destroyListener;
 
@@ -106,14 +108,6 @@ export class AppActionsDeck {
     await modal.present();
   }
 
-  private async openEmbed() {
-    const modal: HTMLIonModalElement = await modalController.create({
-      component: 'app-embed'
-    });
-
-    await modal.present();
-  }
-
   private async openMoreActions($event: UIEvent) {
     if (!$event) {
       return;
@@ -132,15 +126,11 @@ export class AppActionsDeck {
         } else if (detail.data.action === MoreAction.JUMP_TO) {
           await this.openSlideNavigate();
         } else if (detail.data.action === MoreAction.SHARE) {
-          shareStore.state.share = {
-            deck: editorStore.state.deck,
-            userName: userStore.state.name,
-            userSocial: userStore.state.social
-          };
+          share();
         } else if (detail.data.action === MoreAction.PUBLISH) {
           this.actionPublish.emit();
         } else if (detail.data.action === MoreAction.EMBED) {
-          await this.openEmbed();
+          this.openEmbed.emit();
         }
       }
     });
@@ -273,7 +263,7 @@ export class AppActionsDeck {
             <ion-label aria-hidden="true">{i18n.state.editor.present}</ion-label>
           </button>
 
-          <app-action-share class="wider-devices" onOpenEmbed={() => this.openEmbed()}></app-action-share>
+          <app-action-share class="wider-devices"></app-action-share>
 
           <app-action-help class="wider-devices"></app-action-help>