+++ studio/src/app/components/editor/deck/app-share-options/app-share-options.tsx
@@ -17,7 +17,7 @@ export class AppMoreShareOptions {
     return (
       <Host>
         {this.renderUpdate()}
-        {this.renderEmbed()}
+        {editorStore.state.deck !== null && this.renderEmbed()}
         {this.renderShareLink()}
       </Host>
     );