+++ studio/src/global/app.ts
@@ -69,10 +69,10 @@ setupDeckGoConfig({
   },
   cloud: {
     api: {
-      cdn: 'https://unpkg.com/@deckdeckgo/firebase@2.0.0/dist/deckdeckgo-firebase/index.esm.js'
+      cdn: 'https://unpkg.com/@deckdeckgo/firebase@3.0.0/dist/deckdeckgo-firebase/index.esm.js'
     },
     signIn: {
-      cdn: 'https://unpkg.com/@deckdeckgo/firebase@2.0.0/dist/deckdeckgo-firebase/deckdeckgo-firebase.esm.js',
+      cdn: 'https://unpkg.com/@deckdeckgo/firebase@3.0.0/dist/deckdeckgo-firebase/deckdeckgo-firebase.esm.js',
       tag: 'deckgo-firebase-signin'
     }
   }