+++ utils/editor/src/models/data/doc.ts
@@ -1,9 +1,16 @@
+import {Meta} from './meta';
+import {Deploy} from './deploy';
+
 export interface DocData {
   name: string;
 
   owner_id: string | undefined;
 
   paragraphs?: string[];
+
+  meta?: Meta;
+
+  deploy?: Deploy;
 
   created_at?: Date | number | BigInt;
   updated_at?: Date | number | BigInt;